Split phase tests by functionality
